requirements.txt
mss==4.0.3
numpy==1.17.3
opencv-python==4.1.2.30
pypiwin32
